+++ hpeOneView/oneview_client.py
@@ -107,6 +107,7 @@ from hpeOneView.resources.settings.appliance_device_snmp_v3_users import Applian
 from hpeOneView.resources.settings.appliance_node_information import ApplianceNodeInformation
 from hpeOneView.resources.settings.appliance_health_status import ApplianceHealthStatus
 from hpeOneView.resources.settings.appliance_proxy_configuration import ApplianceProxyConfiguration
+from hpeOneView.resources.settings.appliance_network_interfaces import ApplianceNetworkInterfaces
 from hpeOneView.resources.settings.ha_nodes import HANodes
 from hpeOneView.resources.settings.appliance_time_and_locale_configuration import ApplianceTimeAndLocaleConfiguration
 from hpeOneView.resources.settings.versions import Versions
@@ -206,6 +207,7 @@ class OneViewClient(object):
         self.__appliance_node_information = None
         self.__appliance_health_status = None
         self.__appliance_proxy_configuration = None
+        self.__appliance_network_interfaces = None
         self.__versions = None
         self.__backups = None
         self.__login_details = None
@@ -1129,6 +1131,16 @@ class OneViewClient(object):
             ApplianceProxyConfiguration:
         """
         return ApplianceProxyConfiguration(self.__connection)
+
+    @property
+    def appliance_network_interfaces(self):
+        """
+        Gets the ApplianceNetworkInterfaces API client.
+
+        Returns:
+            ApplianceNetworkInterfaces:
+        """
+        return ApplianceNetworkInterfaces(self.__connection)
 
     @property
     def ha_nodes(self):